# kitty-specs/010-account-settings/contracts/openapi.yaml
openapi: 3.0.3
info: { title: Account and Settings API, version: 1.0.0 }
paths:
  /accounts:
    post:
      summary: Create account
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateAccountRequest' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/AccountResponse' } } } }
    get:
      summary: List accounts
      security: [ { bearerAuth: [] } ]
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AccountResponse' } } } }
  /accounts/{id}:
    put:
      summary: Update account
      security: [ { bearerAuth: [] } ]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateAccountRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AccountResponse' } } } }
    delete:
      summary: Delete account
      security: [ { bearerAuth: [] } ]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses:
        '204': { description: No Content }
        '409': { description: Account is referenced by transactions }
  /settings:
    get:
      summary: Get settings
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/SettingsResponse' } } } }
    put:
      summary: Update settings
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateSettingsRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/SettingsResponse' } } } }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    CreateAccountRequest:
      type: object
      required: [name, type, balance]
      properties: { name: { type: string, maxLength: 100 }, type: { type: string, enum: [cash, bank, card] }, balance: { type: number } }
    UpdateAccountRequest:
      type: object
      properties: { name: { type: string, maxLength: 100 }, type: { type: string, enum: [cash, bank, card] }, balance: { type: number } }
    AccountResponse:
      type: object
      required: [id, name, type, balance, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string, enum: [cash, bank, card] }
        balance: { type: number }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UpdateSettingsRequest:
      type: object
      properties:
        firstName: { type: string, maxLength: 80 }
        lastName: { type: string, maxLength: 80 }
        profilePic: { type: string }
        location: { type: string, maxLength: 120 }
        address: { type: string, maxLength: 255 }
    SettingsResponse:
      type: object
      required: [firstName, lastName, profilePic, location, address]
      properties:
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        profilePic: { type: string, nullable: true }
        location: { type: string, nullable: true }
        address: { type: string, nullable: true }